Fixes #204 Made skill cards identcal
